Tiny debugs before RLCT hyperparameter test on GPU
- Unpack error in distill loss block in sampler function
- Early stop threshold type change to float in pydantic model

+++ di_automata/config_setup.py
@@ -483,7 +483,7 @@ class MainConfig(BaseModel):
     num_eval_batches: Optional[int] = Field(default=20)
     early_stop_patience: Optional[int] = Field(default=5, description="Number of evaluation steps with no improvement in log loss before stopping.")
     early_stop_smoothing_window: Optional[int] = Field(default=5, description="Size of moving average window to smooth out loss.")
-    early_stop_acc_threshold: Optional[int] = Field(default=99.9, description="Percent accuracy to immediately stop training at.")
+    early_stop_acc_threshold: Optional[float] = Field(default=99.9, description="Percent accuracy to immediately stop training at.")
     
     # Set by validator
     run_name: Optional[str]
